+++ source/gestalt/application/FirstPersonCamera.cpp
@@ -7,7 +7,7 @@
 
 namespace gestalt::application {
 
-  void FirstPersonCamera::update(float64 delta_seconds, const UserInput& movement,
+  void FirstPersonCamera::update(float32 delta_seconds, const UserInput& movement,
                                  FirstPersonCameraData& data) {
     auto& camera_data = data;
     const auto mouse_pos = glm::vec2(movement.mouse_position_x_rel, movement.mouse_position_y_rel);
